Code review fixes
Refactor file-related utils from EditorUtils to FileUtils. Move readAsText to FileUtils.
