+++ src/DocumentManager.js
@@ -35,7 +35,7 @@ define(function (require, exports, module) {
     var NativeFileSystem    = require("NativeFileSystem").NativeFileSystem,
         ProjectManager      = require("ProjectManager"),
         PreferencesManager  = require("PreferencesManager"),
-        EditorUtils         = require("EditorUtils"),
+        FileUtils           = require("FileUtils"),
         CommandManager      = require("CommandManager"),
         Async               = require("Async"),
         Commands            = require("Commands");
@@ -339,7 +339,7 @@ define(function (require, exports, module) {
     /**
      * Null only of editor is not open yet. If a Document is created on empty text, or text with
      * inconsistent line endings, the Document defaults to the current platform's standard endings.
-     * @type {null|EditorUtils.LINE_ENDINGS_CRLF|EditorUtils.LINE_ENDINGS_LF}
+     * @type {null|FileUtils.LINE_ENDINGS_CRLF|FileUtils.LINE_ENDINGS_LF}
      */
     Document.prototype._lineEndings = null;
 
@@ -369,9 +369,9 @@ define(function (require, exports, module) {
         this.isDirty = false;
         
         // Sniff line-ending style
-        this._lineEndings = EditorUtils.sniffLineEndings(rawText);
+        this._lineEndings = FileUtils.sniffLineEndings(rawText);
         if (!this._lineEndings) {
-            this._lineEndings = EditorUtils.getPlatformLineEndings();
+            this._lineEndings = FileUtils.getPlatformLineEndings();
         }
     };
     
@@ -384,7 +384,7 @@ define(function (require, exports, module) {
         // CodeMirror.getValue() always returns text with LF line endings; fix up to match line
         // endings preferred by the document, if necessary
         var codeMirrorText = this._editor.getValue();
-        if (this._lineEndings === EditorUtils.LINE_ENDINGS_LF) {
+        if (this._lineEndings === FileUtils.LINE_ENDINGS_LF) {
             return codeMirrorText;
         } else {
             return codeMirrorText.replace(/\n/g, "\r\n");
@@ -415,9 +415,9 @@ define(function (require, exports, module) {
         this.diskTimestamp = newTimestamp;
         
         // Re-sniff line-ending style too
-        this._lineEndings = EditorUtils.sniffLineEndings(text);
+        this._lineEndings = FileUtils.sniffLineEndings(text);
         if (!this._lineEndings) {
-            this._lineEndings = EditorUtils.getPlatformLineEndings();
+            this._lineEndings = FileUtils.getPlatformLineEndings();
         }
     };
